Implementation of excercise of the Grunt for week 2.
In this exercise, you will learn to use Grunt, the task runner. You will install Grunt CLI and install Grunt packages using NPM. Thereafter you will configure a Grunt file with a set of tasks to build and serve your web project. At the end of this exercise, you will be able to:

Install Grunt CLI and Grunt packages in your project
Configure a Grunt file with a set of tasks to build a web project from a source, and serve the built project using a server.

Conclusions

In this exercise you have learnt how to configure a Grunt file to perform several tasks. You were able to build a distribution folder and start a server with livereload to serve the web page.