Objectives and Outcomes
In this exercise, you will explore Angular scope and the use of scope as a glue between the view and controller. You will also learn about the ngShow directive. At the end of this exercise, you will be able to:

Understand the use of Angular Scope.
Use scope to connect the view and the controller.
Make use of the ngShow directive.

Conclusions

In this exercise, you learnt more about the use of Angular $Scope. You also learnt about the ngShow directive.